Refactor tensor indexer.
